Define __all__ for expressions module imported with import *
