+++ pysparkling/sql/expressions/aggregate/stat_aggregations.py
@@ -118,3 +118,9 @@ class Kurtosis(SimpleStatAggregation):
 
     def __str__(self):
         return "kurtosis({0})".format(self.column)
+
+
+__all__ = [
+    "Avg", "VarPop", "VarSamp", "Sum", "StddevPop", "StddevSamp",
+    "Skewness", "Min", "Max", "Kurtosis", "Count"
+]
