+++ pysparkling/sql/expressions/dates.py
@@ -433,3 +433,11 @@ class ToUTCTimestamp(Expression):
 
     def __str__(self):
         return "to_utc_timestamp({0}, {1})".format(self.column, self.tz)
+
+
+__all__ = [
+    "ToUTCTimestamp", "FromUTCTimestamp", "TruncTimestamp", "TruncDate", "ParseToDate",
+    "ParseToTimestamp", "UnixTimestamp", "CurrentTimestamp", "FromUnixTime", "WeekOfYear", "NextDay", "MonthsBetween", "LastDay",
+    "DayOfYear", "DayOfMonth", "DayOfWeek", "Month", "Quarter", "Year", "DateDiff", "DateSub", "DateAdd", "DateFormat", "CurrentDate",
+    "AddMonths", "Hour", "Minute", "Second"
+]
