+++ pysparkling/sql/expressions/explodes.py
@@ -67,3 +67,6 @@ class PosExplodeOuter(PosExplode):
 
     def __str__(self):
         return "posexplode_outer"
+
+
+__all__ = ["PosExplodeOuter", "PosExplode", "ExplodeOuter", "Explode"]
