+++ pysparkling/sql/expressions/literals.py
@@ -11,3 +11,6 @@ class Literal(Expression):
 
     def __str__(self):
         return str(self.value)
+
+
+__all__ = ["Literal"]
