+++ pysparkling/sql/internal_utils/writers.py
@@ -5,7 +5,9 @@ import shutil
 
 from pysparkling import Row
 from pysparkling.sql.casts import cast_to_string, get_time_formatter
-from pysparkling.sql.functions import Aggregation, col, StarOperator
+from pysparkling.sql.expressions.aggregate.aggregations import Aggregation
+from pysparkling.sql.expressions.mappers import StarOperator
+from pysparkling.sql.functions import col
 from pysparkling.sql.internal_utils.readwrite import to_option_stored_value
 from pysparkling.sql.utils import AnalysisException
 from pysparkling.utils import portable_hash, get_json_encoder
